pyproject.toml
[project]
name = "askjg-demo-gemini-pcc"
version = "0.1"
description = "Voice AI demo assistant using Gemini Live native audio"
requires-python = ">=3.10"
dependencies = [
    # Pipecat with Gemini Live and Daily transport
    # Note: 'webrtc' extra is for local testing only, not needed in production
    "pipecat-ai[google,daily,silero,runner,webrtc]==0.0.98",
    "pipecatcloud",
    "python-dotenv",
    "loguru",
    # Call recording to GCS
    "gcloud-aio-storage",
    "aiohttp",
    # End-of-call reporting
    "httpx",
    # Post-call processing (summary + transcript)
    "google-genai>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
